Changed int and float to 64 bit variants

+++ src/lib.rs
@@ -7,8 +7,8 @@ pub struct PmlStruct {
 
 enum PmlElem {
     PmlString(String),
-    PmlInt(i32),
-    PmlFloat(f32)
+    PmlInt(i64),
+    PmlFloat(f64)
 }
 
 impl PmlStruct {
@@ -19,14 +19,14 @@ impl PmlStruct {
         }
     }
 
-    pub fn get_int(&self, key: &str) -> &i32 {
+    pub fn get_int(&self, key: &str) -> &i64 {
         match self.elements.get(key) {
             Some(PmlElem::PmlInt(i)) => i,
             _ => panic!("Not an int")
         }
     }
 
-    pub fn get_float(&self, key: &str) -> &f32 {
+    pub fn get_float(&self, key: &str) -> &f64 {
         match self.elements.get(key) {
             Some(PmlElem::PmlFloat(f)) => f,
             _ => panic!("Not a float")
@@ -52,9 +52,9 @@ fn parse_lines(lines: Vec<String>) -> PmlStruct {
     let mut elements: HashMap<String, PmlElem> = HashMap::new();
     for line in lines {
         let (key, value) = line.split_once("=").unwrap();
-        if let Ok(num) = value.parse::<i32>() {
+        if let Ok(num) = value.parse::<i64>() {
             elements.insert(key.to_string(), PmlElem::PmlInt(num));
-        } else if let Ok(num) = value.parse::<f32>() {
+        } else if let Ok(num) = value.parse::<f64>() {
             elements.insert(key.to_string(), PmlElem::PmlFloat(num));
         } else {
             elements.insert(key.to_string(), PmlElem::PmlString(value.to_string()));
